+++ src/messages/Message.cpp
@@ -25,12 +25,12 @@ using namespace literals;
 Message::Message()
     : parseTime(QTime::currentTime())
 {
-    DebugCount::increase("messages");
+    DebugCount::increase(DebugObject::Message);
 }
 
 Message::~Message()
 {
-    DebugCount::decrease("messages");
+    DebugCount::decrease(DebugObject::Message);
 }
 
 ScrollbarHighlight Message::getScrollBarHighlight() const